Save Bazel workspace cache only on specified workflow events. (#15)
This is a mitigation for cache entries that are close to the GitHub Actions cache size limit. In particular, this can avoid the case where the cache entry for the default branch gets evicted by saving a cache entry for a pull request. That case can lead to cache misses, as GitHub actions only restores cache entries from the current branch or the default branch.